google_shopping_agent/requirements.txt
# Google Shopping Agent Dependencies

# HTTP client for SERP API
httpx>=0.25.0

# Supabase Python client
supabase>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Async support
asyncio-throttle>=1.0.0

# Data validation (optional but recommended)
pydantic>=2.0.0

# Logging enhancements (optional)
structlog>=23.0.0

# Streamlit UI
streamlit>=1.29.0

# Data handling
pandas>=2.0.0
openpyxl>=3.1.0  # Excel export
